requirements.txt
aiohttp==3.7.3
altgraph==0.17
async-timeout==3.0.1
attrs==20.3.0
auto-py-to-exe==2.7.11
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2020.12.5
chardet==3.0.4
discord-py-slash-command==1.1.2
discord.py==1.6.0
Eel==0.12.4
fire==0.4.0
future==0.18.2
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
idna==2.10
macholib==1.14
multidict==5.1.0
numpy==1.20.2
openai==0.3.0
optional-django==0.1.0
Pillow==8.2.0
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
python-dotenv==0.15.0
requests==2.25.1
six==1.15.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
whichcraft==0.6.1
yarl==1.6.3
zope.event==4.5.0
zope.interface==5.2.0
